refactor(ui)!: co-locate the modal host and description-list entries with their nodes
The modal host renders modal wire nodes through the openModal effect,
so it moves next to the modal adapter as
components/modal/modal-host.tsx; the entry.* adapters and their shared
semantic context move into components/description-list/. The package
root keeps only node-type-agnostic runtime. Deep imports change from
@lattice-php/ui/modal to @lattice-php/ui/components/modal/modal-host;
the framework runtime re-exports the same names.
