+++ packages/action/resources/js/components/action-form.test.tsx
@@ -5,7 +5,7 @@ import type { Node, Plugin } from "@lattice-php/lattice";
 import { formComponents } from "@lattice-php/form";
 import { useFormContext } from "@lattice-php/form/toolkit";
 import { fakeNode, jsonResponse, renderWithRegistry } from "@lattice-php/core/test-support";
-import { ModalProvider } from "@lattice-php/ui/modal";
+import { ModalProvider } from "@lattice-php/ui/components/modal/modal-host";
 import { actionComponents } from "@lattice-php/action/plugin";
 
 type ValidateFieldsOptions = { onSuccess?: () => void; onValidationError?: () => void };
