+++ packages/framework/resources/js/runtime.ts
@@ -18,7 +18,11 @@ export { useTable } from "@lattice-php/table/hooks/use-table";
 export { useTableSelection } from "@lattice-php/table/hooks/use-table-selection";
 export { getBulkActionNodes } from "@lattice-php/table/lib/bulk";
 export { useAction } from "@lattice-php/action/hooks/use-action";
-export { MODAL_MISSING_ERROR, useEmbeddedModal, useModal } from "@lattice-php/ui/modal";
+export {
+  MODAL_MISSING_ERROR,
+  useEmbeddedModal,
+  useModal,
+} from "@lattice-php/ui/components/modal/modal-host";
 export {
   Badge,
   Button,
