+++ packages/media/resources/js/rich-editor/media-image-dialog.tsx
@@ -2,7 +2,7 @@ import type { Editor } from "@tiptap/core";
 import type { Node as WireNode } from "@lattice-php/core/types";
 import { translate, useT } from "@lattice-php/ui/i18n";
 import { Dialog, DialogContent, DialogHeader } from "@lattice-php/ui/primitives/dialog";
-import { MODAL_MISSING_ERROR, useEmbeddedModal } from "@lattice-php/ui/modal";
+import { MODAL_MISSING_ERROR, useEmbeddedModal } from "@lattice-php/ui/components/modal/modal-host";
 import { LibraryView, type MediaRow } from "../components/library-view";
 
 /**
