+++ packages/ui/resources/js/components/button/button-adapter-modal.test.tsx
@@ -4,7 +4,7 @@ import { createRegistry, eagerComponent } from "@lattice-php/core/registry";
 import { Renderer } from "@lattice-php/core/renderer";
 import { renderWithRegistry, fakeNode, TextProbe } from "@lattice-php/core/test-support";
 import type { Node } from "@lattice-php/core/types";
-import { ModalProvider } from "../../modal";
+import { ModalProvider } from "../modal/modal-host";
 import { useClickBehavior } from "../../click-behavior";
 import ButtonAdapter from "./button-adapter";
 import ModalAdapter from "../modal/modal-adapter";
