+++ packages/ui/resources/js/components/description-list/entries.tsx
