+++ packages/ui/resources/js/components/description-list/entry-row.tsx
