+++ packages/ui/resources/js/components/link/link-adapter-action.test.tsx
@@ -3,7 +3,7 @@ import { beforeEach, describe, expect, it, vi } from "vitest";
 import { ActionInteractionProvider } from "@lattice-php/action";
 import { fakeNode, jsonResponse } from "@lattice-php/core/test-support";
 import type { Node, ComponentPropsOf } from "@lattice-php/core/types";
-import { ModalProvider } from "@lattice-php/ui/modal";
+import { ModalProvider } from "../modal/modal-host";
 import LinkAdapter from "./link-adapter";
 
 const apiFetch = vi.hoisted(() => vi.fn<() => Promise<Response>>());
