+++ packages/ui/resources/js/components/menu-item/menu-item-adapter-action.test.tsx
@@ -3,7 +3,7 @@ import { beforeEach, describe, expect, it, vi } from "vitest";
 import { ActionInteractionProvider } from "@lattice-php/action";
 import { fakeNode } from "@lattice-php/core/test-support";
 import type { Node, ComponentPropsOf } from "@lattice-php/core/types";
-import { ModalProvider } from "../../modal";
+import { ModalProvider } from "../modal/modal-host";
 import MenuItemAdapter from "./menu-item-adapter";
 
 const apiFetch = vi.hoisted(() => vi.fn<() => Promise<Response>>());
