+++ packages/ui/resources/js/components/modal/modal-adapter.browser.test.tsx
@@ -2,7 +2,7 @@ import { describe, expect, it } from "vitest";
 import { createRegistry, eagerComponent } from "@lattice-php/core/registry";
 import { renderWithRegistry } from "@lattice-php/core/browser-test-support";
 import { fakeNode } from "@lattice-php/core/test-support";
-import { ModalProvider, useModal } from "../../modal";
+import { ModalProvider, useModal } from "./modal-host";
 import ModalAdapter from "./modal-adapter";
 
 const registry = createRegistry({
